pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agenkit"
version = "0.1.0"
description = "The foundation layer for AI agents"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "Scott Friedman", email = "scttfrdmn@users.noreply.github.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Typing :: Typed",
]
dependencies = [
    "httpx[http2]>=0.27.0",
    "aiohttp>=3.9.0",
    "aioquic>=1.0.0",
    "websockets>=12.0",
    "grpcio>=1.60.0",
    "protobuf>=4.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "bandit>=1.7.5",
    "pylint>=3.0.0",
    "hypothesis>=6.0.0",
    "grpcio-tools>=1.60.0",
]
benchmarks = [
    "pytest-benchmark>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/agenkit/agenkit"
Documentation = "https://docs.agenkit.dev"
Repository = "https://github.com/agenkit/agenkit"
Issues = "https://github.com/agenkit/agenkit/issues"

[tool.hatch.build.targets.wheel]
packages = ["agenkit", "proto"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",
]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "slow: marks tests as slow running tests (deselect with '-m \"not slow\"')",
    "chaos: marks tests as chaos engineering tests (deselect with '-m \"not chaos\"')",
    "property: marks tests as property-based tests with Hypothesis (deselect with '-m \"not property\"')",
]

[tool.mypy]
strict = true
warn_unreachable = true
warn_unused_configs = true
pretty = true
show_column_numbers = true
show_error_context = true
files = ["agenkit"]

# Less strict for tests and benchmarks
[[tool.mypy.overrides]]
module = ["tests.*", "benchmarks.*"]
disallow_untyped_defs = false
disallow_untyped_calls = false

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "RUF",  # Ruff-specific rules
    "N",    # pep8-naming
    "ASYNC",# flake8-async
    "S",    # bandit security
    "PTH",  # flake8-use-pathlib
    "PL",   # pylint
    "PERF", # perflint
    "FURB", # refurb
    "LOG",  # flake8-logging
    "G",    # flake8-logging-format
]
ignore = [
    "S101",    # Allow assert statements (used in tests)
    "PLR0913", # Too many arguments - sometimes necessary for middleware
    "PLR2004", # Magic value used in comparison - acceptable for tests/examples
    "PLR0912", # Too many branches - acceptable for complex logic
    "PLC0415", # Import not at top - often necessary for circular imports
    "G004",    # f-strings in logging - modern Python, more efficient than %
    "G201",    # .error(exc_info=True) vs .exception() - both valid approaches
    "PERF203", # try-except in loop - necessary for robust error handling
    "PERF401", # Manual list comprehension - sometimes clearer than comprehension
    "PLW2901", # Redefined loop variable - acceptable in specific cases
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",    # Allow assert in tests
    "PLR2004", # Allow magic values in tests
    "S105",    # Possible hardcoded password (test fixtures)
    "S106",    # Possible hardcoded password (test fixtures)
]
"examples/**/*.py" = [
    "S105",   # Possible hardcoded password (examples)
    "S106",   # Possible hardcoded password (examples)
    "T201",   # Print statements allowed in examples
]
"benchmarks/**/*.py" = [
    "S101",    # Allow assert in benchmarks
    "PLR2004", # Allow magic values in benchmarks
]
"agenkit/adapters/python/grpc_server.py" = [
    "N802",   # Function names must be CamelCase for protobuf compatibility
]

[tool.ruff.lint.isort]
known-first-party = ["agenkit"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.coverage.run]
source = ["agenkit"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.bandit]
exclude_dirs = ["tests", "benchmarks", "examples"]
skips = ["B101", "B601"]  # Allow assert, subprocess without shell=True check

[tool.pylint.main]
py-version = "3.10"
ignore = ["tests", "benchmarks", "examples"]
extension-pkg-whitelist = ["grpcio", "protobuf"]

[tool.pylint.messages_control]
disable = [
    "C0103",  # Invalid name - conflicts with established naming
    "C0114",  # Missing module docstring - not all modules need them
    "C0115",  # Missing class docstring - covered by ruff
    "C0116",  # Missing function docstring - covered by ruff
    "R0903",  # Too few public methods - acceptable for simple classes
    "R0913",  # Too many arguments - sometimes necessary
    "W0212",  # Protected access - needed for testing
]

[tool.pylint.format]
max-line-length = 100

[tool.pylint.design]
max-args = 8
max-attributes = 12
